Restaurant Time based Model. Able to do the following:
Calculate how many staffs are redundant or is there any.
Maximum number of customers that can be served in the given targeted time.
Calculate the amount of time required to serve the input number of customers.
Asserts that all customers have been served in Targeted Units of Time.
Input includes number of customers and staff, target total timing, duration for events such as how long a customer takes to pay the bill.